Update README.md
These are top 5 favorites in Movies/Series.

In Movies, it was added my favorite which is Greyhound, Avatar, and Battle of Midway since they are new release in few years. Greyhound shows how they encounter the U-boat. And Battle of Midway shows the history the battle between US and Japan.

In Series, my top 1 favorite is One Piece since every scene reflects the real life situation and it is connected to my life about the story. It is my latest favorite series and it will release more chapters or episodes in the future.